+++ CHANGELOG.md
@@ -2,6 +2,36 @@
 
 All notable changes to this project will be documented in this file. See [standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.
 
+## [18.9.0](https://github.com/lifinance/types/compare/v18.8.0...v18.9.0) (2026-09-15)
+
+
+### Features
+
+* add optional referrer to TransferMetadata ([#566](https://github.com/lifinance/types/issues/566)) ([d8ea4f0](https://github.com/lifinance/types/commit/d8ea4f04733dece23c41fbc87bfd81b78d406064))
+* centrifuge icon ([#581](https://github.com/lifinance/types/issues/581)) ([17b723a](https://github.com/lifinance/types/commit/17b723a6490312834af91296bc1710bbb3e2edac))
+
+## [18.8.0](https://github.com/lifinance/types/compare/v18.7.0...v18.8.0) (2026-09-14)
+
+
+### Features
+
+* add INTENT_REFUND_FAILED substatus for failed Smart Deposits refunds ([#578](https://github.com/lifinance/types/issues/578)) ([635f469](https://github.com/lifinance/types/commit/635f469ca187b418d9f6fc10f342ab21c649eb32))
+* add LI.FI Earn API types ([#561](https://github.com/lifinance/types/issues/561)) ([9bf5da1](https://github.com/lifinance/types/commit/9bf5da10adc6c24cea84779ace0bc3faad145e27))
+
+## [18.7.0](https://github.com/lifinance/types/compare/v18.6.0...v18.7.0) (2026-09-11)
+
+
+### Features
+
+* add per-side liquidity source parameters ([#576](https://github.com/lifinance/types/issues/576)) ([24b60e1](https://github.com/lifinance/types/commit/24b60e1faeb1ed97e72a8a001718b250a07d21b4))
+
+## [18.6.0](https://github.com/lifinance/types/compare/v18.5.0...v18.6.0) (2026-09-10)
+
+
+### Features
+
+* add PermitSingle to TypedDataPrimaryTypes ([#577](https://github.com/lifinance/types/issues/577)) ([f3759a2](https://github.com/lifinance/types/commit/f3759a299b7d3fa415d359d94a3f523a116b1ddd))
+
 ## [18.5.0](https://github.com/lifinance/types/compare/v18.4.0...v18.5.0) (2026-09-08)